+++ src/index.ts
@@ -11,11 +11,13 @@ import { createWatchedProxy } from './utils/proxy';
 import {
   BrandingOptionsConfig,
   CaptionsOptions,
+  ConnectToFramePayload,
   EmbeddedInstance,
   FeatureFlag,
   InitialRoomSettings,
   InitOptions,
   InstanceProperties,
+  QueuedEventListener,
   ReceiveMessage,
   RoomJoinedPayload,
   SendMessage,
@@ -52,6 +54,8 @@ export class DigitalSambaEmbedded extends EventEmitter implements EmbeddedInstan
   stored: Stored;
 
   permissionManager = new PermissionManager(this);
+
+  queuedEventListeners: QueuedEventListener[] = [];
 
   constructor(
     options: Partial<InitOptions> = {},
@@ -176,6 +180,49 @@ export class DigitalSambaEmbedded extends EventEmitter implements EmbeddedInstan
       } else {
         this._emit(message.type, message);
       }
+    }
+  };
+
+  addFrameEventListener = (
+    eventName: string,
+    target: 'document' | 'window',
+    listener: (...args: any[]) => void
+  ) => {
+    const customEventName = `frameEvent_${eventName}_${target}`;
+
+    if (this.connected) {
+      if (!this.listenerCount(customEventName)) {
+        this.sendMessage({ type: 'connectEventListener', data: { eventName, target } });
+      }
+    } else {
+      this.queuedEventListeners.push({
+        operation: 'connectEventListener',
+        event: eventName,
+        target,
+      });
+    }
+
+    this.on(customEventName, listener);
+  };
+
+  removeFrameEventListener = (
+    eventName: string,
+    target: 'document' | 'window',
+    listener: (...args: any[]) => void
+  ) => {
+    const customEventName = `frameEvent_${eventName}_${target}`;
+    this.off(customEventName, listener);
+
+    if (this.connected) {
+      if (!this.listenerCount(eventName)) {
+        this.sendMessage({ type: 'disconnectEventListener', data: { eventName, target } });
+      }
+    } else {
+      this.queuedEventListeners.push({
+        operation: 'disconnectEventListener',
+        event: eventName,
+        target,
+      });
     }
   };
 
@@ -343,6 +390,13 @@ export class DigitalSambaEmbedded extends EventEmitter implements EmbeddedInstan
         this._emit('roomJoined', { type: 'roomJoined' });
         break;
       }
+      case 'documentEvent': {
+        const { eventName, target, payload } = message.data as any;
+
+        const customEventName = `frameEvent_${eventName}_${target}`;
+
+        this._emit(customEventName, JSON.parse(payload));
+      }
       default: {
         break;
       }
@@ -410,7 +464,12 @@ export class DigitalSambaEmbedded extends EventEmitter implements EmbeddedInstan
   };
 
   private async checkTarget() {
-    this.sendMessage({ type: 'connect', data: this.roomSettings || {} });
+    const payload: ConnectToFramePayload = {
+      ...this.roomSettings,
+      eventListeners: this.queuedEventListeners,
+    };
+
+    this.sendMessage({ type: 'connect', data: payload });
 
     const confirmationTimeout = window.setTimeout(() => {
       this.logError(UNKNOWN_TARGET);
@@ -418,6 +477,8 @@ export class DigitalSambaEmbedded extends EventEmitter implements EmbeddedInstan
 
     this.on('connected', () => {
       this.connected = true;
+      this.queuedEventListeners = [];
+
       clearTimeout(confirmationTimeout);
     });
   }